added useGetList
