+++ src/components/ListOfNotes/index.js
@@ -1,1 +1,2 @@
-export { default } from './ListOfNotes';
+export { default as ListOfNotes } from './ListOfNotes';
+export { default as useGetList } from './useGetList';
