+++ src/components/index.js
@@ -1,2 +1,2 @@
 export { default as Editor } from './Editor';
-export { default as ListOfNotes } from './ListOfNotes';
+export * from './ListOfNotes';
